backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
bcrypt
python-multipart
pydantic
openpyxl
reportlab
Pillow
easyocr
opencv-python-headless
python-dotenv
